+++ CHANGES.rst
@@ -17,9 +17,15 @@ Unreleased
   fields :attr:`~UpdatedFeed.new` and :attr:`~UpdatedFeed.modified`
   became optional.
   (:issue:`96`)
-* Fix bug in :mod:`~reader.plugins.entry_dedupe` that failed updates
-  of feeds with multiple *new* duplicates of the same issue.
+* Fix bug in :mod:`~reader.plugins.entry_dedupe` causing updates to fail
+  if there were multiple *new* duplicates of the same issue.
   (:issue:`292`)
+* Fix bug in :mod:`~reader.plugins.readtime`
+  and :mod:`~reader.plugins.mark_as_read` causing updates to fail
+  if an entry was deleted by another plugin.
+  (:issue:`292`)
+* Fix bug in :mod:`~reader.plugins.mark_as_read` causing updates to fail
+  if an entry had no title.
 * In the CLI, don't suppress the traceback of :exc:`ReaderError`,
   since it would also suppress it for bugs.
 * In the CLI, stop using deprecated :func:`click.get_terminal_size`.